azure-test-project/pyproject.toml
[project]
name = "pinecone-byoc"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pulumi-pinecone-byoc[azure] @ file:///Users/benny/Projects/pulumi-pinecone"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
